transactions_jsapi 微信支付 统一下单

+++ rust/src/src/wx/pay_transactions_jsapi/pay_transactions_jsapi_model.rs
@@ -1,5 +1,6 @@
 use async_graphql::SimpleObject;
 use serde::{Serialize, Deserialize};
+use rust_decimal::Decimal;
 
 #[derive(
   SimpleObject,
@@ -35,7 +36,7 @@ pub struct TransactionsJsapiInput {
   
   pub description: String,
   
-  pub amount: u64,
+  pub amount: Decimal,
   
   /// 【交易结束时间】 订单失效时间，遵循rfc3339标准格式，格式为yyyy-MM-DDTHH:mm:ss+TIMEZONE，yyyy-MM-DD表示年月日，T出现在字符串中，表示time元素的开头，HH:mm:ss表示时分秒，TIMEZONE表示时区（+08:00表示东八区时间，领先UTC8小时，即北京时间）。例如：2015-05-20T13:29:35+08:00表示，北京时间2015年5月20日13点29分35秒。
   pub time_expire: Option<String>,
